+++ CHANGELOG.md
@@ -11,6 +11,6 @@
     01. Added `dev.orne.qdsl.AssigmentTranslator` interface.
     01. Added `dev.orne.qdsl.TranslateVisitor` class.
     01. Added `dev.orne.qdsl.AbstractPathTranslator` class.
-    01. Added `dev.orne.qdsl.PropertyPathTranslator` class.
+    01. Added `dev.orne.qdsl.SimplePathTranslator` class.
     01. Added `dev.orne.qdsl.DelegatedTranslateVisitor` class.
     01. Added `dev.orne.qdsl.Translator` class.